pyproject.toml
[tool.poetry]
name = "chekhovagent"
version = "0.1.0"
description = ""
authors = ["Bazavod <eugeny.bazavod@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
fastapi = "^0.124.2"
uvicorn = "^0.38.0"
python-dotenv = "^1.2.1"
httpx = "^0.28.1"
aiogram = "^3.23.0"
pytest = "^9.0.2"
google-api-python-client = "^2.187.0"
google-auth = "^2.43.0"
google-auth-httplib2 = "^0.2.1"
logging = "^0.4.9.6"
openai = "^2.11.0"
pydantic-settings = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
